+++ homework/vladimir_popov/Task10/Part1.py
@@ -1,0 +1,16 @@
+def finish_me(func):
+
+    def finish_function(text):
+        func(text)
+        print("finished")
+        return func
+
+    return finish_function
+
+
+@finish_me
+def example(text):
+    print(text)
+
+
+example('print me')
