+++ homework/vladimir_popov/Task10/Part3.py
