+++ homework/vladimir_popov/Task10/Part4.py
