+++ homework/vladimir_popov/Task2/file.txt
