+++ homework/vladimir_popov/Task3/Part1.py
