+++ homework/vladimir_popov/Task3/Part2.py
