+++ homework/vladimir_popov/Task3/Part3.py
