+++ homework/vladimir_popov/Task3/Part4.py
