--- homework/vladimir_popov/Task4/Part1.py
@@ -1,4 +1,0 @@
-
-person = ['John', 'Doe', 'New York', '+1372829383739', 'US']
-
-name, last_name, city, phone, country = person
