+++ homework/vladimir_popov/Task4/Task3.py
