+++ homework/vladimir_popov/Task5/Part1.py
@@ -1,15 +1,4 @@
-text = "Etiam tincidunt neque erat, quis molestie enim imperdiet vel.\
-        Integer urna nisl, facilisis vitae semper at, dignissim vitae libero"
 
-new_text = []
+person = ['John', 'Doe', 'New York', '+1372829383739', 'US']
 
-for word in text.split():
-    if "," in word:
-        word = word.replace(",", "ing,")
-    elif "." in word:
-        word = word.replace(".", "ing.")
-    else:
-        word += "ing"
-    new_text.append(word)
-
-print(" ".join(new_text))
+name, last_name, city, phone, country = person
