+++ homework/vladimir_popov/Task5/Part2.py
@@ -1,10 +1,14 @@
 
-for i in range(1, 101):
-    if i % 3 == 0 and i % 5 == 0:
-        print("FuzzBuzz")
-    elif i % 3 == 0:
-        print("Fuzz")
-    elif i % 5 == 0:
-        print("Buzz")
-    else:
-        print(i)
+string_1 = "результат операции: 42"
+string_2 = "результат операции: 514"
+string_3 = "результат работы программы: 9"
+
+colon_1_index, colon_2_index, colon_3_index = string_1.index(":"), string_2.index(":"), string_3.index(":")
+
+digit_1 = int(string_1[colon_1_index + 1:]) + 10
+digit_2 = int(string_2[colon_2_index + 1:]) + 10
+digit_3 = int(string_3[colon_3_index + 1:]) + 10
+
+print(f"{string_1[:colon_1_index + 1]} {digit_1}")
+print(f"{string_2[:colon_2_index + 1]} {digit_2}")
+print(f"{string_3[:colon_3_index + 1]} {digit_3}")
