+++ homework/vladimir_popov/Task5/Part3.py
