+++ homework/vladimir_popov/Task6/Part2.py
@@ -1,9 +1,10 @@
 
-words = {'I': 3, 'love': 5, 'Python': 1, '!': 50}
-
-for key, value in words.items():
-    i = 0
-    while i < value:
-        print(key, end="")
-        i += 1
-    print()
+for i in range(1, 101):
+    if i % 3 == 0 and i % 5 == 0:
+        print("FuzzBuzz")
+    elif i % 3 == 0:
+        print("Fuzz")
+    elif i % 5 == 0:
+        print("Buzz")
+    else:
+        print(i)
