+++ homework/vladimir_popov/Task7/Part3.py
