+++ homework/vladimir_popov/Task8/Part1.py
@@ -1,15 +1,10 @@
+from random import random, randint, choice
 
-temperatures = [
-    20, 15, 32, 34, 21, 19, 25, 27, 30, 32, 34, 30, 29, 25, 27, 22, 22, 23, 25, 29,
-    29, 31, 33, 31, 30, 32, 30, 28, 24, 23
-]
+salary = int(input("Введите зарплату: "))
 
-hot_days = list(filter(lambda x: x > 28, temperatures))
+bonus = choice([True, False])
 
-hottest_day = max(hot_days)
-average_day = round(sum(hot_days) / len(hot_days))
-coldest_day = min(hot_days)
-
-print(hottest_day)
-print(average_day)
-print(coldest_day)
+if bonus:
+    print(f"${salary + randint(0, 500)}")
+else:
+    print(f"${salary}")
